Refactor somewhat
